feat(proof): refine host rework — split config, unify errors, reorganize kv
