+++ crates/proof/host/Cargo.toml
@@ -39,12 +39,12 @@ alloy-eips.workspace = true
 alloy-network.workspace = true
 base-protocol.workspace = true
 base-alloy-network.workspace = true
-base-proof-std-fpvm.workspace = true
+
 base-consensus-providers.workspace = true
 alloy-rpc-types-beacon.workspace = true
 base-proof = { workspace = true, features = ["std"] }
 alloy-rpc-types = { workspace = true, features = ["debug"] }
-clap = { workspace = true, features = ["derive", "env"] }
+
 base-consensus-genesis = { workspace = true, features = ["serde", "std"] }
 base-alloy-rpc-types-engine = { workspace = true, features = ["serde", "std"] }
