+++ crates/proof/host/src/backend/mod.rs
@@ -3,6 +3,3 @@ pub use offline::OfflineHostBackend;
 
 mod online;
 pub use online::OnlineHostBackend;
-
-mod util;
-pub use util::store_ordered_trie;
