+++ crates/proof/host/src/lib.rs
@@ -3,33 +3,30 @@
 mod error;
 pub use error::{HostError, Result};
 
-mod eth;
-pub use eth::rpc_provider;
+mod config;
+pub use config::{HostConfig, HostProviders};
+
+mod host;
+pub use host::Host;
 
 mod server;
-pub use server::{PreimageServer, PreimageServerError};
+pub use server::PreimageServer;
+
+mod handler;
+pub use handler::{handle_hint, parse_blob_hint};
 
 mod kv;
 #[cfg(feature = "disk")]
 pub use kv::DiskKeyValueStore;
-pub use kv::{KeyValueStore, MemoryKeyValueStore, SharedKeyValueStore, SplitKeyValueStore};
+pub use kv::{
+    BootKeyValueStore, KeyValueStore, MemoryKeyValueStore, SharedKeyValueStore, SplitKeyValueStore,
+    store_ordered_trie,
+};
 
 mod backend;
-pub use backend::{OfflineHostBackend, OnlineHostBackend, store_ordered_trie};
+pub use backend::{OfflineHostBackend, OnlineHostBackend};
 
 #[cfg(feature = "precompiles")]
 mod precompiles;
 #[cfg(feature = "precompiles")]
 pub use precompiles::execute;
-
-mod host;
-pub use host::Host;
-
-mod providers;
-pub use providers::HostProviders;
-
-mod handler;
-pub use handler::{handle_hint, parse_blob_hint};
-
-mod local_kv;
-pub use local_kv::LocalInputs;
