+++ crates/proof/host/src/server.rs
@@ -1,5 +1,3 @@
-//! This module contains the [`PreimageServer`] struct and its implementation.
-
 use std::sync::Arc;
 
 use base_proof_preimage::{
@@ -8,30 +6,15 @@ use base_proof_preimage::{
 use tokio::spawn;
 use tracing::{error, info};
 
+use crate::HostError;
+
 /// The [`PreimageServer`] is responsible for waiting for incoming preimage requests and
 /// serving them to the client.
 #[derive(Debug)]
 pub struct PreimageServer<P, H, B> {
-    /// The oracle server.
     oracle_server: P,
-    /// The hint reader.
     hint_reader: H,
-    /// [`PreimageServerBackend`] that routes hints and retrieves preimages.
     backend: Arc<B>,
-}
-
-/// An error that can occur when handling preimage requests
-#[derive(Debug, thiserror::Error)]
-pub enum PreimageServerError {
-    /// A preimage request error.
-    #[error("Failed to serve preimage request: {0}")]
-    PreimageRequestFailed(PreimageOracleError),
-    /// An error when failed to serve route hint.
-    #[error("Failed to route hint: {0}")]
-    RouteHintFailed(PreimageOracleError),
-    /// Task failed to execute to completion.
-    #[error("Join error: {0}")]
-    ExecutionError(#[from] tokio::task::JoinError),
 }
 
 impl<P, H, B> PreimageServer<P, H, B>
@@ -40,31 +23,25 @@ where
     H: HintReaderServer + Send + Sync + 'static,
     B: PreimageServerBackend + Send + Sync + 'static,
 {
-    /// Create a new [`PreimageServer`] with the given [`PreimageOracleServer`],
-    /// [`HintReaderServer`], and [`PreimageServerBackend`].
+    /// Creates a new [`PreimageServer`].
     pub const fn new(oracle_server: P, hint_reader: H, backend: Arc<B>) -> Self {
         Self { oracle_server, hint_reader, backend }
     }
 
     /// Starts the [`PreimageServer`] and waits for incoming requests.
-    pub async fn start(self) -> Result<(), PreimageServerError> {
+    pub async fn start(self) -> Result<(), HostError> {
         let server =
             spawn(Self::start_oracle_server(self.oracle_server, Arc::clone(&self.backend)));
         let hint_router =
             spawn(Self::start_hint_router(self.hint_reader, Arc::clone(&self.backend)));
 
         tokio::select! {
-            s = server => s?,
-            h = hint_router => h?,
+            s = server => s.map_err(|e| HostError::Custom(e.to_string()))?,
+            h = hint_router => h.map_err(|e| HostError::Custom(e.to_string()))?,
         }
     }
 
-    /// Starts the oracle server, which waits for incoming preimage requests and serves them to the
-    /// client.
-    async fn start_oracle_server(
-        oracle_server: P,
-        backend: Arc<B>,
-    ) -> Result<(), PreimageServerError> {
+    async fn start_oracle_server(oracle_server: P, backend: Arc<B>) -> Result<(), HostError> {
         info!(target: "host_server", "starting oracle server");
         loop {
             match oracle_server.next_preimage_request(backend.as_ref()).await {
@@ -72,15 +49,13 @@ where
                 Err(PreimageOracleError::IOError(_)) => return Ok(()),
                 Err(e) => {
                     error!(target: "host_server", error = %e, "failed to serve preimage request");
-                    return Err(PreimageServerError::PreimageRequestFailed(e));
+                    return Err(HostError::PreimageRequestFailed(e));
                 }
             }
         }
     }
 
-    /// Starts the hint router, which waits for incoming hints and routes them to the appropriate
-    /// handler.
-    async fn start_hint_router(hint_reader: H, backend: Arc<B>) -> Result<(), PreimageServerError> {
+    async fn start_hint_router(hint_reader: H, backend: Arc<B>) -> Result<(), HostError> {
         info!(target: "host_server", "starting hint router");
         loop {
             match hint_reader.next_hint(backend.as_ref()).await {
@@ -88,7 +63,7 @@ where
                 Err(PreimageOracleError::IOError(_)) => return Ok(()),
                 Err(e) => {
                     error!(target: "host_server", error = %e, "failed to route hint");
-                    return Err(PreimageServerError::RouteHintFailed(e));
+                    return Err(HostError::RouteHintFailed(e));
                 }
             }
         }
